KERI-12: Big rename of 8-02-1
Renamed, what was 8-02, to 8-02-1 to match the instructor's
demonstration-performance student lab.

Tests pass locally and the Makefile rule is successful

Changes to be committed:
	modified:   code/Makefile
	renamed:    code/include/8-02-linked_list.h -> code/include/8-02-1-linked_list.h
	modified:   code/include/8-02-sort_functions.h
	renamed:    code/src/8-02-linked_list-main.c -> code/src/8-02-1-linked_list-main.c
	renamed:    code/src/8-02-linked_list-solution.c -> code/src/8-02-1-linked_list-solution.c
